+++ RELEASE_NOTES.md
@@ -20,8 +20,8 @@ This version adds the very first (and so far limited) support for `SwiftUI`, as 
 There are some new views that can be used in SwiftUI-based apps and keyboard extensions:
 
 * The new `ImageKeyboardButton` view lets you bind an `.image` action to a SwiftUI button.
-* The new `KeyboardActionGrid` distributes keyboard actions evenly within a grid.
-* The new `KeyboardActionGridRow` is used by this grid, for each row in the grid.
+* The new `KeyboardGrid` distributes actions evenly within a grid.
+* The new `KeyboardGridRow` is used for each row in the grid.
 * The new `NextKeyboardButton` sets itself up with a `globe` icon and works as a standard "next keyboard" button.
 
 * The new `.globe` `Image` extension returns the icon that is used for "next keyboard".